move envVariableFeatureFlags to sourceHandler
